Update README.md
This Python module defines an Order class that facilitates the comparison between two orders based on their prices. It uses Python's special (or "dunder") method __gt__ to override the greater-than operator (>), making it possible to easily compare two instances of the class in terms of their price attribute. This module can be a handy addition to applications dealing with e-commerce, inventory management, or anywhere order comparisons are necessary.